+++ src/mask.cpp
@@ -139,16 +139,16 @@ int GEMMI_MAIN(int argc, char **argv) {
         std::fprintf(stderr, "Grid: %d x %d x %d\n", g.nu, g.nv, g.nw);
         std::fprintf(stderr, "Spacing: %.3f, %.3f, %.3f\n",
                              g.spacing[0], g.spacing[1], g.spacing[2]);
-        int np = g.data.size();
+        size_t np = g.data.size();
         double vol = st.cell.volume;
-        std::fprintf(stderr, "Total points: %d\n", np);
+        std::fprintf(stderr, "Total points: %zu\n", np);
         std::fprintf(stderr, "Unit cell volume: %.1f A^3\n", vol);
         std::fprintf(stderr, "Volume per point: %.3f A^3\n", vol / np);
         if (g.spacegroup) {
           std::fprintf(stderr, "Spacegroup: %s\n", g.spacegroup->hm);
           int na = g.spacegroup->operations().order();
           std::fprintf(stderr, "ASU volume: %.1f A^3\n", vol / na);
-          std::fprintf(stderr, "Points per ASU: %d\n", np / na);
+          std::fprintf(stderr, "Points per ASU: %.1f\n", double(np) / na);
         } else {
           std::fprintf(stderr, "No spacegroup\n");
         }
@@ -160,13 +160,13 @@ int GEMMI_MAIN(int argc, char **argv) {
           for (const gemmi::Atom& atom : res.atoms)
             mask.grid.set_points_around(atom.pos, radius, 1);
       if (p.options[Verbose]) {
-        int n = std::count(mask.grid.data.begin(), mask.grid.data.end(), 1);
-        std::fprintf(stderr, "Points masked by model: %d\n", n);
+        size_t n = std::count(mask.grid.data.begin(), mask.grid.data.end(), 1);
+        std::fprintf(stderr, "Points masked by model: %zu\n", n);
       }
       mask.grid.symmetrize_max();
       if (p.options[Verbose]) {
-        int n = std::count(mask.grid.data.begin(), mask.grid.data.end(), 1);
-        std::fprintf(stderr, "After symmetrizing: %d\n", n);
+        size_t n = std::count(mask.grid.data.begin(), mask.grid.data.end(), 1);
+        std::fprintf(stderr, "After symmetrizing: %zu\n", n);
       }
       mask.update_ccp4_header(0, true);
       mask.write_ccp4_map(output);
